// src/types/models/gprDisc.ts
export interface GprDisc {
  rectangleLineNumber: Number;
  profileType: String;
  crackProfileNumber: Number;
  typeOfCrack: String;
  typeOfDisc: String;
  dip: Number;
  dipDirection: Number;
  mapReferenceSystem: Number;
  startingVertexX: Number;
  startingVertexY: Number;
  startingVertexZ: Number;
  endVertexX: Number;
  endVertexY: Number;
  endVertexZ: Number;
  nX: Number;
  nY: Number;
  nZ: Number;
}